Development Environment Changes
Got rid of the "Build" directory as you can store builds in the "extra"
folder as substitute. Added the Binaries folder with the readme file
on what to place in there. Also, updated README.md with some extra
information on what Maelstrom is as well as what the "extra" folder
is really for