Updater Changes
The sckCheckUpdate and tmrCheckUpdate controls are no longer used.
Instead an MSXML2.XMLHTTP object is used to get update information.
The Code Speak Distribution System will be used in place of
querying a single file for the program's latest version as well as
a page on the Code Speak website to display update information
such as changes since the installed version and a download link
to the latest version