Treat "Product must be downgraded" as a false positive
This is a reversal from a decision over 2 years ago. Back then
I thought the issue had to do with certain keys, however I was
wrong. I'm not quite sure why this is an issue but recently had
done some research. On top of an updated VersionCheck.dll (thanks
Hdx for the updated code and BNETDocs for at least considering the
research I had done), this issue should be less noticeable than it
used to be. I've noticed there are still some issues but have
since decided to treat this as a false positive.